requirements.txt
# Core
requests>=2.28
PyYAML>=6.0
pandas>=2.0
tqdm>=4.65
python-dotenv>=1.0
rapidfuzz>=2.13
unidecode>=1.3
python-dateutil>=2.8
regex>=2023.8.8

# Excel export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# PDF extraction
PyMuPDF>=1.23.0      # fitz for PDF text
pdfplumber>=0.10.0

# HTML parsing for web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Firecrawl for intelligent web scraping
firecrawl-py>=0.0.5

# Gemini AI
google-generativeai>=0.3.0

# For BibTeX export
bibtexparser>=1.2.0

# Rate limiting
ratelimit>=2.2.1
